Stop using make when running Desktop reftests.
This moves the logic that was previously in Make either into the
runtests.py file, or into the mach command. In general the latter
is only used for setting defaults which a user may reasonably want
to override.

This enables some code simplification as we can use the same
command line arguments parsing code for mach and runtests.py
and permits the removal of lots of options preprocessing
code from mach_commands.py.